cleanup redundant UI code
